Fixed bytestride source for uint

+++ src/main/java/io/github/chadj2/mesh/buffer/UIntLineIndices.java
@@ -92,6 +92,6 @@ public class UIntLineIndices extends ScalarBuffer<Integer> {
 
   @Override
   protected void setBufferViewByteStride(BufferView _bufferView) {
-    _bufferView.setByteStride(Byte.BYTES);
+    _bufferView.setByteStride(Integer.BYTES);
   }
 }
